Refactor header-nav
Introduce @xo-union/utilities and refactor all previous references to
utilities
