+++ packages/pattern-library/utilities/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@xo-union/utilities",
+  "version": "0.1.0",
+  "main": "index.js",
+  "dependencies": {
+    "css-module-builder": "^1.0.0-beta.5",
+    "@xo-union/colors": "^0.1.0"
+  }
+}
