+++ pattern-library/docs/articles/doc-components.md
@@ -1,7 +1,7 @@
 ---
 $imports:
   '{ Snippet, Demo, PropTypesTable, InstallSnippet }': '#docs/doc-components'
-  utils: '@xo-union/bootstrap/utilities'
+  spacing: '@xo-union/utilities/spacing'
   PropTypesTableMetadata: '!!react-docgen-loader!#docs/doc-components/PropTypesTable'
   SnippetsMetadata:       '!!react-docgen-loader!#docs/doc-components/Snippet'
   DemoMetadata:           '!!react-docgen-loader!#docs/doc-components/Demo'
@@ -12,14 +12,14 @@ $imports:
 
 > These are components used to build this documentation
 
-<div className={utils.mt5}></div>
+<div className={spacing.mt5}></div>
 
 ## PropTypesTable Component
 
 <span>{PropTypesTableMetadata.description}</span>
 
 
-<div className={utils.mt4}></div>
+<div className={spacing.mt4}></div>
 
 #### Properties and Demo
 
@@ -29,13 +29,13 @@ $imports:
   <PropTypesTable metadata={PropTypesTableMetadata.props} exclude={["default"]}/>
 </Demo>
 
-<div className={utils.mt5}></div>
+<div className={spacing.mt5}></div>
 
 ## Snippet Component
 
 <span>{SnippetsMetadata.description}</span>
 
-<div className={utils.mt4}></div>
+<div className={spacing.mt4}></div>
 
 #### Demo
 
@@ -45,19 +45,19 @@ $imports:
   </Snippet>
 </Demo>
 
-<div className={utils.mt4}></div>
+<div className={spacing.mt4}></div>
 
 #### Properties
 
 <PropTypesTable metadata={SnippetsMetadata.props} exclude={["default"]} />
 
-<div className={utils.mt5}></div>
+<div className={spacing.mt5}></div>
 
 ## Demo Component
 
 <span>{DemoMetadata.description}</span>
 
-<div className={utils.mt4}></div>
+<div className={spacing.mt4}></div>
 
 #### Demo
 
@@ -71,29 +71,27 @@ Hello Demo
   </Demo>
 </Demo>
 
-<div className={utils.mt4}></div>
+<div className={spacing.mt4}></div>
 
 #### Properties
 
 <PropTypesTable metadata={DemoMetadata.props} exclude={["default"]} />
 
-<div className={utils.mt5}></div>
+<div className={spacing.mt5}></div>
 
 ## InstallSnippet Component
 
 <span>{InstallSnippetMetadata.description}</span>
 
-<div className={utils.mt4}></div>
+<div className={spacing.mt4}></div>
 
 #### Demo
-
-This is a demo of the Demo component, we added a pink border to make the representation clearer.
 
 <Demo>
   <InstallSnippet packageJson={{name: "test-package", version: "1.0.0"}} />
 </Demo>
 
-<div className={utils.mt4}></div>
+<div className={spacing.mt4}></div>
 
 #### Properties
