+++ pattern-library/docs/src/components/Layout/styles.css
@@ -1,4 +1,4 @@
 .article {
   composes: col from '@xo-union/bootstrap/grid';
-  composes: py-4 from '@xo-union/bootstrap/utilities';
+  composes: py-4 from '@xo-union/utilities/spacing';
 }
