+++ pattern-library/docs/src/demos/Form/styles.css
@@ -1,11 +1,11 @@
 @value ( white, jay-lighter-90 ) from '@xo-union/colors';
 
 .demo-for-gray-theme {
-  composes: py3 from '@xo-union/bootstrap/utilities';
-  background-color: white;
+  composes: py3 from '@xo-union/utilities/spacing';
+  composes: white-bg from '@xo-union/colors/utilities';
 }
 
 .demo-for-white-theme {
-  composes: py3 from '@xo-union/bootstrap/utilities';
-  background-color: jay-lighter-90;
+  composes: py3 from '@xo-union/utilities/spacing';
+  composes: jay-lighter-90-bg from '@xo-union/colors/utilities';
 }
