+++ pattern-library/src/bootstrap/entrypoints.json
@@ -3,6 +3,5 @@
   "nav.scss",
   "navbar.scss",
   "reboot.scss",
-  "tables.scss",
-  "utilities.scss"
+  "tables.scss"
 ]
