+++ pattern-library/src/buttons/css/index.scss
@@ -14,7 +14,7 @@
 @value baby-min-width: initial;
 
 .btn-transitions {
-  transition: all .2s ease-out;
+  transition: all .1s ease-out;
   transition-property: background-color, color, border-color;
 }
