+++ pattern-library/src/colors/entrypoints.json
@@ -1,7 +1,4 @@
 [
-  {
-    "from": "classes.scss",
-    "to": "css/index.cssm"
-  },
-  "variables.scss"
+  "variables.scss",
+  "utilities.scss"
 ]
