+++ pattern-library/src/fields/css/index.scss
@@ -1,5 +1,5 @@
 /* Static classes */
-@value ( container, field-container ) from './styles/container';
+@value ( container ) from './styles/container';
 @value ( field-label, invalid-field, valid-field, requirements ) from './styles/field';
 @value ( dropdown-caret ) from './styles/dropdown-field';
 @value ( dropdown-item, dropdown-item-check ) from './styles/dropdown-item';
