+++ pattern-library/src/fields/css/styles/container.scss
@@ -1,12 +1,3 @@
 .container {
   position: relative;
 }
-
-.field-container {
-  composes: text-left from '@xo-union/bootstrap/utilities';
-  composes: container;
-
-  * {
-    box-sizing: border-box;
-  }
-}
