+++ pattern-library/src/typography/utilities.scss
@@ -1,2 +1,2 @@
 @import '~#/_shared/bootstrap-overrides';
-@import '~bootstrap/scss/utilities';
+@import '~bootstrap/scss/utilities/text';
