+++ pattern-library/src/utilities/display.scss
@@ -1,0 +1,2 @@
+@import '~#/_shared/bootstrap-overrides';
+@import '~bootstrap/scss/utilities/display';
