+++ pattern-library/src/utilities/effects.scss
@@ -1,0 +1,5 @@
+@value ( background-gray-8 ) from '@xo-union/colors';
+
+.light-shadow {
+  box-shadow: 0 .125rem .188rem 0 background-gray-8;
+}
