+++ pattern-library/src/utilities/entrypoints.json
@@ -1,0 +1,8 @@
+[
+  "effects.scss",
+  "display.scss",
+  "flexbox.scss",
+  "spacing.scss",
+  "vertical-align.scss",
+  "toggleable.scss"
+]
